Use Country entity from neo4j module in COS API
