--- packages/server/customer-os-api/entity/country.go
@@ -1,22 +1,0 @@
-package entity
-
-import (
-	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-neo4j-repository/neo4jutil"
-)
-
-type CountryEntity struct {
-	Name      string
-	CodeA2    string
-	CodeA3    string
-	PhoneCode string
-
-	DataloaderKey string
-}
-
-type CountryEntities []CountryEntity
-
-func (CountryEntity) Labels(string) []string {
-	return []string{
-		neo4jutil.NodeLabelCountry,
-	}
-}
