+++ packages/server/customer-os-api/mapper/country_mapper.go
@@ -1,8 +1,8 @@
 package mapper
 
 import (
-	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-api/entity"
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-api/graph/model"
+	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-neo4j-repository/entity"
 )
 
 func MapEntityToCountry(entity *entity.CountryEntity) *model.Country {
